Fix split link of dbms/programs/odbc-bridge
